Moved signin code to google page so it can be reused even when the CH signin button is not clicked
